project/requirements.txt
langchain
langchain_community
langchain_google_genai
python-dotenv
pgvector
transformers
requests
